change transport connect signature
The transport needs to create either a read/write stream, or a write-only
stream. The latter is used when handling messages from a non-peer stream. Future
messages from the same address should be delivered to the non-peer stream, until
the cluster configuration is changed and a proper peer is created.