test & improve Candidate
Don't bind handleMessage when creating the consumer so it can be stubbed in the
tests.

Send RejectEntries when receiving an AppendEntries message with an older term.
Presumably the candidate has already requested a vote from the former leader
with the newer term, but it doesn't hurt to send another message.