test & improve Follower
Don't bind handleMessage when creating the consumer so it can be stubbed in the
tests.

Rename flag used to ignore the next election timeout for improved clarity.

Call message handlers with the term in its own parameter, for consistency with
the Candidate and Leader roles. This makes it easier to share test code for this
mapping which occurs in all three classes.