emit role-change events synchronously
Change the emitter to emit events synchronously. Any errors are rethrown
asynchronously so not to affect Raft behavior. If a listener crashes subsequent
listeners for the same emission are not invoked.

The role-change events are emitted immediately after the role is started, unless
starting the role immediately causes it to change. In other words, when the
event is emitted the role is active.