Simplify how leadership is claimed
Appending a Noop requires data to be persisted, which means it'll take longer
for the leader to actually claim leadership. Instead send a heartbeat.

The Noop approach was taken so that the leader can commit entries from previous
terms before new data is appended. This is still a worthwhile goal, but a) only
needs to happen if there are indeed uncommitted entries from previous terms, and
b) shouldn't delay the leadership claim.

Restoring the Noop behavior is left as a TODO.