Ensure follower skips next election even if rejecting entries
When the follower receives an AppendEntries message, even if the entries are
rejected, skip the next election. The leader is clearly still active.

(Note that the election is not skipped if the leader's term is behind. In that
case it's already been deposed as a leader but it just doesn't know it yet.)