Optimize follower convergence
Set the conflicting index included in the RejectEntries response to be the
lowest index at which the follower should converge. E.g. if the follower's log
is empty, send `1` so the leader knows to send its first entry. If the log is
behind the received previous log index, send the index the log is at now. If the
log is ahead (but conflicted), send the previous log index as received from the
leader.

This allows the follower to converge with the leader much more quickly, which is
good for fuzz testing, as otherwise too much time is spent on the convergence
alone. This is also an optimization mentioned in the Raft paper.